feat: adiciona na rota de atualizar home as prescricoes
